+++ test/unit/index.spec.ts
@@ -1,6 +1,11 @@
 import { expect, expectTypeOf } from "@infra-blocks/test";
 import { isNumber, isObjectNotNull } from "@infra-blocks/types";
-import { findCause, findCauseByType } from "../../src/index.js";
+import {
+  findCause,
+  findCauseByType,
+  hasCause,
+  hasCauseByType,
+} from "../../src/index.js";
 
 describe("index", () => {
   describe(findCause.name, () => {
@@ -90,6 +95,35 @@ describe("index", () => {
       const result = findCauseByType(root, CustomError);
       expectTypeOf(result).toEqualTypeOf<CustomError | undefined>();
       expect(result).to.equal(leaf);
+    });
+  });
+  describe(hasCause.name, () => {
+    it("should return false when the predicate does not match any error in the causal chain", () => {
+      const leaf = new Error("leaf");
+      const mid = new Error("trunk", { cause: leaf });
+      const root = new Error("root", { cause: mid });
+      expect(hasCause(root, (e) => e === "not any error")).to.be.false;
+    });
+    it("should return true when the predicate matches an error in the causal chain", () => {
+      const leaf = new Error("leaf");
+      const mid = new Error("trunk", { cause: leaf });
+      const root = new Error("root", { cause: mid });
+      expect(hasCause(root, (e) => e instanceof Error && e.message === "leaf"))
+        .to.be.true;
+    });
+  });
+  describe(hasCauseByType.name, () => {
+    it("should return false when no instances can be found in the causal chain", () => {
+      const leaf = new Error("leaf");
+      const mid = new TypeError("trunk", { cause: leaf });
+      const root = new RangeError("root", { cause: mid });
+      expect(hasCauseByType(root, SyntaxError)).to.be.false;
+    });
+    it("should return true when an instance can be found in the causal chain", () => {
+      const leaf = new SyntaxError("that's not how you spell it mfka!");
+      const mid = new TypeError("trunk", { cause: leaf });
+      const root = new RangeError("root", { cause: mid });
+      expect(hasCauseByType(root, SyntaxError)).to.be.true;
     });
   });
 });